--- src/components/SuperButton/SuperButton.css
@@ -1,97 +1,0 @@
-/* Общие стили для всех кнопок */
-.button {
-    padding: 10px 20px;
-    border: 1px solid transparent;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-    transition: background-color 0.3s, color 0.3s, border 0.3s;
-}
-
-/* Кнопка для светлой темы (белая тема — кнопка черная) */
-.button_theme_white {
-    background-color: #000000;  /* Черный фон */
-    color: white;               /* Белый текст */
-    border: 1px solid #000000;
-}
-
-.button_theme_white:hover {
-    background-color: #3C3B3BFF;  /* Темно-серый при наведении */
-}
-.button_theme_white.button_hovered {
-    background-color: #3C3B3BFF;  /* Темно-серый при наведении */
-}
-/* Кнопка для темной темы (темная тема — кнопка белая) */
-.button_theme_black {
-    background-color: #ffffff;  /* Белый фон */
-    color: black;             /* Черный текст */
-    border: 1px solid #ffffff;
-}
-
-.button_theme_black:hover {
-    background-color: #B8B7B7FF;  /* Светло-серый при наведении */
-}
-.button_theme_black.button_hovered {
-    background-color: #B8B7B7FF;  /* Светло-серый при наведении */
-}
-
-
-
-
-
-/* Второстепенная кнопка (secondary) для темной темы */
-.button_theme_black.button_isSecondary {
-    background-color: transparent;
-    color: #f3f3f3;
-    border: 1px solid #ffffff;
-}
-
-.button_theme_black.button_isSecondary.button_disabled {
-    background-color: #161616;  /* Светло-серый фон */
-    color: #5f5f5f;             /* Серый текст */
-    border: 1px solid #5f5f5f;  /* Светлая рамка */
-    cursor: not-allowed;        /* Курсор не доступен */
-}
-
-.button_theme_black.button_isSecondary:hover {
-    background-color: rgba(255, 255, 255, 0.2); /* Легкое осветление */
-}
-.button_theme_black.button_isSecondary.button_hovered {
-    background-color: rgba(255, 255, 255, 0.2); /* Легкое осветление */
-}
-
-
-/* Второстепенная кнопка (secondary) для светлой темы */
-.button_theme_white.button_isSecondary {
-    background-color: transparent;
-    color: rgba(2, 2, 2, 0.9);
-    border: 1px solid #000000;
-}
-
-.button_theme_white.button_isSecondary:hover {
-    background-color: rgba(0, 0, 0, 0.2); /* Легкое затемнение */
-}
-.button_theme_white.button_isSecondary.button_hovered {
-    background-color: rgba(0, 0, 0, 0.2); /* Легкое затемнение */
-    border: 1px solid #cfcfcf;
-}
-.button_hovered {
-    background-color: rgba(0, 0, 0, 0.2); /* Легкое затемнение */
-}
-
-.button_theme_white.button_isSecondary.button_disabled {
-    background-color: #f4f3f3;  /* Светло-серый фон */
-    color: #a0a0a0;             /* Серый текст */
-    border: 1px solid #d0d0d0;  /* Светлая рамка */
-    cursor: not-allowed;        /* Курсор не доступен */
-}
-
-
-
-/* Неактивная кнопка */
-.button_disabled {
-    background-color: #f4f3f3;  /* Светло-серый фон */
-    color: #a0a0a0;             /* Серый текст */
-    border: 1px solid #d0d0d0;  /* Светлая рамка */
-    cursor: not-allowed;        /* Курсор не доступен */
-}
